% Auto-generated: do not edit by hand
\name{button}

\alias{button}

\title{Button component}

\description{

}

\usage{
button(children=NULL, id=NULL, style=NULL, className=NULL,
disabled=NULL, kind=NULL, size=NULL, n_clicks=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. The children of the element}

\item{id}{Character. }

\item{style}{Named list. }

\item{className}{Character. }

\item{disabled}{Logical. }

\item{kind}{Character. }

\item{size}{A value equal to: 'default', 'field', 'small', 'sm', 'lg', 'xl'. }

\item{n_clicks}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

