% Auto-generated: do not edit by hand
\name{column}

\alias{column}

\title{Column component}

\description{

}

\usage{
column(children=NULL, columnSizes=NULL, offsetSizes=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }

\item{columnSizes}{List of characters. }

\item{offsetSizes}{List of characters. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

