% Auto-generated: do not edit by hand
\name{datePickerInput}

\alias{datePickerInput}

\title{DatePickerInput component}

\description{

}

\usage{
datePickerInput(id=NULL, disabled=NULL, hideLabel=NULL,
iconDescription=NULL, invalid=NULL, invalidText=NULL,
labelText=NULL, placeholder=NULL, size=NULL, warn=NULL,
warnText=NULL, value=NULL, datePickerType=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{disabled}{Logical. Specify whether or not the input should be disabled}

\item{hideLabel}{Logical. Specify if the label should be hidden}

\item{iconDescription}{Character. The description of the calendar icon.}

\item{invalid}{Logical. Specify whether or not the input should be invalid}

\item{invalidText}{Character. Specify the text to be rendered when the input is invalid}

\item{labelText}{Character. Provide the text that will be read by a screen reader when visiting this control}

\item{placeholder}{Character. Specify the placeholder text}

\item{size}{A value equal to: 'sm', 'xl'. Specify the size of the Date Picker Input. Currently supports either sm or xl as an option.}

\item{warn}{Logical. Specify whether the control is currently in warning state}

\item{warnText}{Character. Provide the text that is displayed when the control is in warning state}

\item{value}{Character. Current value of the date}

\item{datePickerType}{A value equal to: 'range', 'single', 'simple'. The type of the date picker}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

