% Auto-generated: do not edit by hand
\name{slider}

\alias{slider}

\title{Slider component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
slider(id=NULL, name=NULL, inputType=NULL, ariaLabelInput=NULL,
labelText=NULL, disabled=NULL, light=NULL,
hideTextInput=NULL, value=NULL, min=NULL, max=NULL,
step=NULL, stepMultiplier=NULL)
}

\arguments{
\item{id}{Character. Slider id}

\item{name}{Character. Form item name}

\item{inputType}{Character. The form element type}

\item{ariaLabelInput}{Character. The ARIA label for the <input>}

\item{labelText}{Character. Label text}

\item{disabled}{Logical. Disabled}

\item{light}{Logical. Light variant}

\item{hideTextInput}{Logical. Without text input}

\item{value}{Numeric. The value}

\item{min}{Numeric. The minimum value}

\item{max}{Numeric. The maximum value}

\item{step}{Numeric. The step}

\item{stepMultiplier}{Numeric. The step factor for Shift+arrow keys}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

