% Auto-generated: do not edit by hand
\name{tab}

\alias{tab}

\title{Tab component}

\description{

}

\usage{
tab(children=NULL, className=NULL, disabled=NULL, label=NULL,
role=NULL, value=NULL)
}

\arguments{
\item{children}{A value equal to: proptypes.arrayof(proptypes.node), proptypes.node. Tab content}

\item{className}{Character. Specify an optional className to be added to your Tab}

\item{disabled}{Logical. Whether your Tab is disabled.}

\item{label}{Character. Provide the contents of your Tab}

\item{role}{Character. Provide an accessibility role for your Tab}

\item{value}{Character. Value of the tab}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

