% Auto-generated: do not edit by hand
\name{uIShell}

\alias{uIShell}

\title{UIShell component}

\description{
UIShell is a default layout with the header and a sidebar
}

\usage{
uIShell(children=NULL, name=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. Pages}

\item{name}{Character. Plataform Name}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

