% Auto-generated: do not edit by hand
\name{checkbox}

\alias{checkbox}

\title{Checkbox component}

\description{

}

\usage{
checkbox(id=NULL, label=NULL, disabled=NULL, value=NULL, title=NULL,
style=NULL, className=NULL)
}

\arguments{
\item{id}{Character. }

\item{label}{Character. }

\item{disabled}{Logical. }

\item{value}{Logical. }

\item{title}{Character. }

\item{style}{Named list. }

\item{className}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

