% Auto-generated: do not edit by hand
\name{checkbox}

\alias{checkbox}

\title{Checkbox component}

\description{
Checkbox Input
}

\usage{
checkbox(id=NULL, label=NULL, hideLabel=NULL, disabled=NULL,
value=NULL, title=NULL, style=NULL, className=NULL)
}

\arguments{
\item{id}{Character. Provide an `id` to uniquely identify the Checkbox input}

\item{label}{A list of or a singular dash component, string or number. Provide a label to provide a description of the Checkbox input that you are
exposing to the user}

\item{hideLabel}{Logical. Specify whether the label should be hidden, or not}

\item{disabled}{Logical. Specify whether the Checkbox should be disabled}

\item{value}{Logical. Specify whether the underlying input is checked}

\item{title}{Character. Specify a title for the <label> node for the Checkbox}

\item{style}{Named list. The inline styles}

\item{className}{Character. Specify an optional className to be applied to the <label> node}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

