% Auto-generated: do not edit by hand
\name{link}

\alias{link}

\title{Link component}

\description{

}

\usage{
link(children=NULL, id=NULL, style=NULL, className=NULL,
disabled=NULL, inline=NULL, size=NULL, n_clicks=NULL,
visited=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{style}{Named list. }

\item{className}{Character. }

\item{disabled}{Logical. }

\item{inline}{Logical. }

\item{size}{A value equal to: 'sm', 'lg'. }

\item{n_clicks}{Numeric. }

\item{visited}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

