% Auto-generated: do not edit by hand
\name{radioButtonGroup}

\alias{radioButtonGroup}

\title{RadioButtonGroup component}

\description{

}

\usage{
radioButtonGroup(legendText=NULL, name=NULL, radiosButtons=NULL, id=NULL,
value=NULL, labelPosition=NULL, orientation=NULL,
disabled=NULL)
}

\arguments{
\item{legendText}{Character. }

\item{name}{Character. }

\item{radiosButtons}{List of lists containing elements 'labeltext', 'value', 'id'.
those elements have the following types:
  - labeltext (character; optional)
  - value (character; optional)
  - id (character; optional)s. }

\item{id}{Character. }

\item{value}{Character. }

\item{labelPosition}{A value equal to: 'left', 'right'. }

\item{orientation}{A value equal to: 'horizontal', 'vertical'. }

\item{disabled}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

