% Auto-generated: do not edit by hand
\name{row}

\alias{row}

\title{Row component}

\description{
Grid Row
}

\usage{
row(children=NULL, style=NULL, id=NULL, className=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. The children of the element}

\item{style}{Named list. The inline styles}

\item{id}{Character. The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{className}{Character. The class of the element}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

