% Auto-generated: do not edit by hand
\name{uIPage}

\alias{uIPage}

\title{UIPage component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
uIPage(children=NULL, name=NULL)
}

\arguments{
\item{children}{List of a list of or a singular dash component, string or numbers | a list of or a singular dash component, string or number. Children}

\item{name}{Character. Page name}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

